♻️ refactor: 확장가능성이 적은 interface 제거 및 패키지 구조 변경
