--- src/main/java/utils/Parser.java
@@ -1,5 +1,0 @@
-package utils;
-
-public interface Parser<T> {
-    T parse(String input);
-}
