+++ src/main/java/utils/parser/QueryStringParser.java
@@ -1,9 +1,9 @@
-package utils;
+package utils.parser;
 
-public class QueryStringParser implements Parser {
+public class QueryStringParser{
     private static final String QUERYSTRING_SEPARATOR = "?";
 
-    public String parse(String path) {
+    public static String parse(String path) {
 
         return path.substring(path.indexOf(QUERYSTRING_SEPARATOR) + 1);
     }
