+++ src/test/java/utils/parser/WebProtocolParserTest.java
@@ -1,15 +1,14 @@
-package utils;
+package utils.parser;
 
 import model.WebProtocol;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import utils.parser.WebProtocolParser;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
 
 class WebProtocolParserTest {
-    WebProtocolParser webProtocolParser = new WebProtocolParser();
-
     @Test
     @DisplayName("requestLine을 method, path(request-target), protocol, version으로 분리 할 수 있다.")
     void parsingTest() {
@@ -19,7 +18,7 @@ class WebProtocolParserTest {
         String protocolVersion = "1.1";
 
         //when
-        WebProtocol webProtocolParseResult = webProtocolParser.parse(protocolWithVersion);
+        WebProtocol webProtocolParseResult = WebProtocolParser.parse(protocolWithVersion);
 
         //then
         assertAll(
